add casl

+++ app/api/books/[id]/route.ts
@@ -1,42 +1,43 @@
-/**
- *
- *  GET id
- *  PUT id
- *  DELETE id
- *
- *  patch id admin approve
- *
- *
- */
-
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 import { bookSchema } from "@/lib/zodSchemas";
+import { ability } from "@/lib/casl";
 
 export async function GET(
   req: Request,
   { params }: { params: { id: string } }
 ) {
   try {
-    const books = await prisma.book.findMany({
+    const book = await prisma.book.findMany({
       where: {
         id: params.id,
       },
     });
 
-    //FIXME:   !book return error 400
+    if(!book){
+
+       
+        return NextResponse.json({
+          data: {
+            error: true,
+            message: "no book found",
+            status: 404,
+          },
+        });
+    }
+
 
     return NextResponse.json({
       data: {
         error: false,
-        books,
+        book,
       },
     });
   } catch (error) {
     return NextResponse.json({
       data: {
         error: true,
-        message: "An error occurred while fetching books.",
+        message: "An error occurred while fetching book.",
         status: 500,
       },
     });
@@ -51,6 +52,20 @@ export async function PUT(
     const parsed = bookSchema.safeParse(await req.json());
 
     //FIXME:   !owner Id from req
+    const authUser = JSON.parse(req.headers.get('user') as string)
+
+    if(!authUser || !ability(authUser).can('update','Book')){
+
+       
+        return NextResponse.json({
+          data: {
+            error: true,
+            message: "Unauthenticated",
+            status: 401,
+          },
+        });
+    }
+
 
     if (!parsed.success) {
       return NextResponse.json(
@@ -64,7 +79,7 @@ export async function PUT(
     const books = await prisma.book.update({
       where: {
         id: params.id,
-        ownerId: "clzjwjsny0000xwfcq2s7fw9y",
+        ownerId: authUser?.user?.id,
       },
       data: { title, author, categoryId, price, quantity },
     });
@@ -90,12 +105,27 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   try {
-    //FIXME:   !owner Id from req
+    
+
+    const authUser = JSON.parse(req.headers.get('user') as string)
+
+    if(!authUser || !ability(authUser).can('delete','Book')){
+
+       
+        return NextResponse.json({
+          data: {
+            error: true,
+            message: "Unauthenticated",
+            status: 401,
+          },
+        });
+    }
+
 
     await prisma.book.delete({
       where: {
         id: params.id,
-        ownerId: "clzjwjsny0000xwfcq2s7fw9y",
+        ownerId: authUser?.user?.id ,
       },
     });
